+++ libnavui-androidauto/src/main/java/com/mapbox/androidauto/MapboxCarApp.kt
@@ -1,15 +1,12 @@
 package com.mapbox.androidauto
 
 import android.app.Application
-import com.mapbox.androidauto.configuration.CarAppConfigOwner
-import com.mapbox.androidauto.datastore.CarAppDataStoreOwner
-import com.mapbox.androidauto.navigation.audioguidance.MapboxAudioGuidance
-import com.mapbox.androidauto.navigation.audioguidance.impl.MapboxAudioGuidanceImpl
-import com.mapbox.androidauto.navigation.audioguidance.impl.MapboxAudioGuidanceServicesImpl
 import com.mapbox.androidauto.navigation.location.CarAppLocation
 import com.mapbox.androidauto.navigation.location.impl.CarAppLocationImpl
 import com.mapbox.navigation.base.ExperimentalPreviewMapboxNavigationAPI
 import com.mapbox.navigation.core.lifecycle.MapboxNavigationApp
+import com.mapbox.navigation.ui.app.internal.SharedApp
+import com.mapbox.navigation.ui.voice.internal.MapboxAudioGuidance
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
 
@@ -25,16 +22,6 @@ object MapboxCarApp {
      * Attach observers to the CarAppState to determine which view to show.
      */
     val carAppState: StateFlow<CarAppState> = carAppStateFlow
-
-    /**
-     * Stores preferences that can be remembered across app launches.
-     */
-    val carAppDataStore by lazy { CarAppDataStoreOwner() }
-
-    /**
-     * Attach observers to monitor the configuration of the app and car.
-     */
-    val carAppConfig: CarAppConfigOwner by lazy { CarAppConfigOwner() }
 
     /**
      * Location service available to the car and app.
@@ -62,17 +49,13 @@ object MapboxCarApp {
      */
     fun setup(
         application: Application,
-        audioGuidance: MapboxAudioGuidance = MapboxAudioGuidanceImpl(
-            MapboxAudioGuidanceServicesImpl(),
-            carAppDataStore,
-            carAppConfig
-        ),
-        carAppLocation: CarAppLocation = CarAppLocationImpl(),
+        audioGuidance: MapboxAudioGuidance? = null,
+        carAppLocation: CarAppLocation? = null
     ) {
-        carAppDataStore.setup(application)
-        carAppConfig.setup(application)
+        SharedApp.setup(application, audioGuidance)
 
-        MapboxNavigationApp.registerObserver(audioGuidance)
-        MapboxNavigationApp.registerObserver(carAppLocation)
+        MapboxNavigationApp.registerObserver(
+            carAppLocation ?: CarAppLocationImpl()
+        )
     }
 }